# Speedup Lineage Extraction 