--- a/pkg/pipeline/lineage.go
+++ b/pkg/pipeline/lineage.go
@@ -17,13 +17,16 @@ type sqlParser interface {
 type LineageExtractor struct {
 	sqlParser sqlParser
 	renderer  *jinja.Renderer
+	processed map[string]bool
 }
 
 // NewLineageExtractor creates a new LineageExtractor instance.
 func NewLineageExtractor(parser sqlParser) *LineageExtractor {
+
 	return &LineageExtractor{
 		sqlParser: parser,
 		renderer:  jinja.NewRendererWithYesterday("lineage-parser", "lineage-parser"),
+		processed: make(map[string]bool),
 	}
 }
 
@@ -43,6 +46,12 @@ func (p *LineageExtractor) ColumnLineage(foundPipeline *Pipeline, asset *Asset, 
 	if asset == nil {
 		return nil
 	}
+	key := fmt.Sprintf("%s-%s", foundPipeline.Name, asset.Name)
+	if p.processed[key] {
+		return nil
+	}
+
+	p.processed[asset.Name] = true
 
 	for _, upstream := range asset.Upstreams {
 		upstreamAsset := foundPipeline.GetAssetByName(upstream.Value)